Add bare bones card creation
Mostly just getting a feel for the state update and re-render cycle.

The React tutorial has a form component that deletes the value of the
text box on submit. If the submission fails, this means your value is
lost. I implemented a less user hostile version that continuously saves
the value of the textbox to a "draft" area in the App's state object.