Add an index for sprint related cards
In order to get the cards for a sprint I needed to add a custom index to
the PouchDB so I defined the view function in the sprint model file and
added code to Database.js to load it and create a design doc in the
database. This keeps any app specific code out of Database.js.