Add actions to model/database
To implement the action of assigning a card to a sprint I made a pair of
functions that update the the value associated with the 'sprint' leu in
a card document. I added these functions to the sprint model and updated
Database.js to load and partially apply them so they have access to the
PouchDB instance created at runtime. This keeps app specific logic out
of Database.js.